Extracted DateField from boundDateField (#1186)
* Extracted DateField from boundDateField

* Rename prop
